Moved logic to BalancesManager, added partial historic view
